+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 * make error messages a bit more clear
 * improvements to documentation
 * added more CLI options to control verbosity, user agent, etc.
+* if module cannot be found, search current directory
 
 0.6.0 - 2021-04-12
 ------------------