+++ docs/changelog.rst
@@ -1,11 +1,14 @@
 Changelog
 =========
 
-0.7.0 -
+0.7.0 - 2021-06-04
 ------------------
 
-* add scout command
-
+* add ``spatula scout`` command
+* make error messages a bit more clear
+* improvements to documentation
+* added more CLI options to control verbosity, user agent, etc.
+* if module cannot be found, search current directory
 
 0.6.0 - 2021-04-12
 ------------------